requirements.txt
flask
dnspython
dkimpy
pyspf
scikit-learn
joblib
nltk